requirements.txt
joblib==1.5.3
numpy==2.4.4
pandas==3.0.3
pillow==12.2.0
scikit-learn==1.8.0
scipy==1.17.1
streamlit==1.57.0
